requirements.txt
anyio==4.8.0
Cerberus==1.3.7
certifi==2025.1.31
charset-normalizer==3.4.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
packaging==24.2
playwright==1.50.0
pluggy==1.5.0
pyee==12.1.1
pytest==8.3.4
pytest-asyncio==0.25.3
pytest-base-url==2.1.0
python-dotenv==1.0.1
python-slugify==8.0.4
python-telegram-bot==21.10
PyYAML==6.0.2
sniffio==1.3.1
telegram-logging==1.0.0
text-unidecode==1.3
typing_extensions==4.12.2
urllib3==2.3.0
